chore: extract `execution_result.rs` from `executor.rs`
