+++ crates/executor/src/lib.rs
@@ -8,12 +8,15 @@
 //! Reth executor executes transaction in block of data.
 
 use async_trait::async_trait;
-use executor::ExecutionResult;
+use execution_result::ExecutionResult;
 use reth_interfaces::executor::Error;
 use reth_primitives::{Address, Block};
 
 pub mod config;
 pub mod eth_dao_fork;
+
+/// Execution result types
+pub mod execution_result;
 /// Executor
 pub mod executor;
 /// Wrapper around revm database and types
