+++ crates/stages/src/stages/execution.rs
@@ -10,7 +10,7 @@ use reth_db::{
     transaction::{DbTx, DbTxMut},
 };
 use reth_executor::{
-    executor::AccountChangeSet,
+    execution_result::AccountChangeSet,
     revm_wrap::{State, SubState},
 };
 use reth_primitives::{
